Extract SportsEvent model and update views
Move SportsEvent into its own SportsEvent.swift with sampleEvents, remove HomeView.swift, and update ListView and MapView to use the shared model. ListView now uses a NavigationStack and a scrollable event list with a reusable eventRow; MapView was refactored to display an interactive Map with annotations, selectable event cards, a detail panel, and local join/region controls. These changes centralize the event model and improve map/list UI and interactions.
